Backend/TODO/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
pydantic==2.1.1
pydantic_core==2.4.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
